Complete work on rationalising plot package.
Complete a first draft of the v2 changes document
