+++ pynbody/plot/stars.py
@@ -546,20 +546,8 @@ def sfh(sim, massform=True, trange=None, bins=100, **kwargs):
 	plt.xlabel('Time [Gyr]', fontsize='large')
 	plt.ylabel(r'SFR [M$_\odot$ yr$^{-1}$]', fontsize='large')
 
-	x0, x1 = plt.gca().get_xlim()
-
-
-	# add a z axis on top if it has not been already done by an earlier plot:
-	old_axis = pyplot.gca()
-
-	pz = plt.twiny()
-	labelzs = np.arange(5, int(sim.properties['z']) - 1, -1)
-	times = cosmology.age(sim, labelzs, unit='Gyr')
-	pz.set_xticks(times)
-	pz.set_xticklabels([str(x) for x in labelzs])
-	pz.set_xlim(x0, x1)
-	pz.set_xlabel('$z$')
-	pyplot.sca(old_axis)
+	from .util import add_redshift_axis
+	add_redshift_axis(sim)
 
 	return array.SimArray(thebins, "Gyr"), array.SimArray(sfhist, "Msol yr**-1")
