+++ pynbody/plot/util.py
@@ -1,69 +1,97 @@
 """
-
-util
-====
-
 Utility functions for the plotting module
 
 """
 
+import matplotlib.patches
+import matplotlib.quiver
+import matplotlib.transforms
 import numpy as np
 import scipy as sp
 import scipy.signal
 import scipy.sparse
+from matplotlib import pyplot as plt
 
+from ..analysis import cosmology
+
+
+def add_redshift_axis(sim, labelzs=[0.0, 0.2, 0.5, 1.0, 1.5, 2.0, 3.0, 5.0]):
+    """Add a top redshift axis to a plot with time on the x-axis
+
+    Parameters
+    ----------
+
+    sim : :class:`pynbody.snapshot.SimSnap`
+        The simulation snapshot to which the plot refers
+
+    labelzs : list
+        List of redshifts at which to place labels
+
+    """
+
+    old_axis = plt.gca()
+    x0, x1 = plt.gca().get_xlim()
+    pz = plt.twiny()
+    times = cosmology.age(sim, labelzs, unit='Gyr')
+    pz.set_xticks(times)
+    pz.set_xticklabels([str(x) for x in labelzs])
+    pz.set_xlim(x0, x1)
+    pz.set_xlabel('$z$')
+    plt.sca(old_axis)
 
 def fast_kde(x, y, kern_nx=None, kern_ny=None, gridsize=(100, 100),
              extents=None, nocorrelation=False, weights=None, norm = False, **kwargs):
-    """
-    A faster gaussian kernel density estimate (KDE).  Intended for
-    computing the KDE on a regular grid (different use case than
-    scipy's original scipy.stats.kde.gaussian_kde()).
-
-    Author: Joe Kington
-    License:  MIT License <http://www.opensource.org/licenses/mit-license.php>
-
-    Performs a gaussian kernel density estimate over a regular grid using a
-    convolution of the gaussian kernel with a 2D histogram of the data.
+    """Gaussian kernel density estimation (KDE)
 
     This function is typically several orders of magnitude faster than
     scipy.stats.kde.gaussian_kde for large (>1e7) numbers of points and
-    produces an essentially identical result.
+    produces an essentially identical result. Unlike the scipy original, however,
+    it is limited to using a regular grid.
 
-    **Input**:
+    Parameters
+    ----------
 
-        *x*: array
-            The x-coords of the input data points
+    x : array
+        The x-coords of the input data points
 
-        *y*: array
-            The y-coords of the input data points
+    y : array
+        The y-coords of the input data points
 
-        *kern_nx*: float
-            size (in units of *x*) of the kernel
+    kern_nx : float
+        size (in units of x) of the kernel. If None, the size is determined
+        automatically based on the Scott's factor.
 
-        *kern_ny*: float
-            size (in units of *y*) of the kernel
+    kern_ny : float
+        size (in units of y) of the kernel. If None, the size is determined
+        automatically based on the Scott's factor.
 
-        *gridsize*: (Nx , Ny) tuple (default: 200x200)
-            Size of the output grid
+    gridsize : tuple
+        Size of the output grid (default 100x100)
 
-        *extents*: (default: extent of input data) A (xmin, xmax, ymin, ymax)
-            tuple of the extents of output grid
+    extents : tuple
+        Extents of the output grid as (xmin, xmax, ymin, ymax). Default: extent of input data
 
-        *nocorrelation*: (default: False) If True, the correlation between the
-            x and y coords will be ignored when preforming the KDE.
+    nocorrelation : bool
+        If True, the correlation between the x and y coords will be ignored when
+        preforming the KDE.
 
-        *weights*: (default: None) An array of the same shape as x & y that
-            weighs each sample (x_i, y_i) by each value in weights (w_i).
-            Defaults to an array of ones the same size as x & y.
+    weights : array
+        An array of the same shape as x & y that weighs each sample (x_i, y_i) by each
+        value in weights (w_i). Defaults to an array of ones the same size as x & y.
 
-        *norm*: boolean (default: False)
-            If False, the output is only corrected for the kernel. If True,
-            the result is normalized such that the integral over the area
-            yields 1.
+    norm : bool
+        If False, the output is only corrected for the kernel. If True, the result is
+        normalized such that the integral over the area yields 1.
 
-    **Output**:
-        A gridded 2D kernel density estimate of the input points.
+    Returns
+    -------
+
+    A gridded 2D kernel density estimate of the input points.
+
+
+    :Authors: Joe Kington
+    :License: MIT License <http://www.opensource.org/licenses/mit-license.php>
+
     """
 
     #---- Setup --------------------------------------------------------------
@@ -171,35 +199,115 @@ def fast_kde(x, y, kern_nx=None, kern_ny=None, gridsize=(100, 100),
     return grid
 
 
-def inv_fourier(p, nmin=1000, mmin=1, mmax=7, nphi=100):
-    """
 
-    Invert a profile with fourier coefficients to yield an overdensity
-    map.
 
-    **Inputs:**
 
-    *p* : a :func:`~pynbody.analysis.profile.Profile` object
+def _val_or_rc(val, rc_key):
+    """Return val if it is not None, otherwise return the value of rc_key in matplotlib.rcParams.
 
-    **Optional Keywords:**
+    This is available in some versions of matplotlib, but was added in mid-2023, so we should support
+    older versions for now"""
 
-    *nmin* (1000) : minimum number of particles required per bin
+    return val if val is not None else matplotlib.rcParams[rc_key]
 
-    *mmin* (1)    : lowest multiplicity Fourier component
+class PynbodyQuiverKey(matplotlib.quiver.QuiverKey):
+    """An improved version of matplotlib's QuiverKey, allowing a background color to be specified."""
 
-    *mmax* (7)    : highest multiplicity Fourier component
+    def __init__(self, *args, **kwargs):
+        """An improved quiver key implementation.
 
-    *nphi* (100)  : number of azimuthal bins to use for the map
+        In addition to the arguments of matplotlib.quiver.QuiverKey, additional parameters are below.
 
-    """
+        Parameters
+        ----------
+        boxfacecolor : str or None
+            The background color of the key. If None, the default legend face color is used.
+        boxedgecolor : str or None
+            The edge color of the key. If None, the default legend edge color is used.
+        fancybox : bool or None
+            If True, the box is drawn with a fancy box style. If None, the default legend fancybox style is used.
+            If False, a square-cornered box is drawn.
+        *args:
+            Additional arguments for matplotlib.quiver.QuiverKey
+        **kwargs:
+            Additional keyword arguments for matplotlib.quiver.QuiverKey
 
-    phi_hist = np.zeros((len(p['rbins']), nphi))
-    phi = np.linspace(-np.pi, np.pi, nphi)
-    rbins = p['rbins']
+        """
+        self.boxfacecolor = _val_or_rc(kwargs.pop('boxfacecolor', None),
+                                                     'legend.facecolor')
+        self.boxedgecolor = _val_or_rc(kwargs.pop('boxedgecolor', None),
+                                                  'legend.edgecolor')
 
-    for i in range(len(rbins)):
-        if p['n'][i] > nmin:
-            for m in range(mmin, mmax):
-                phi_hist[i, :] = phi_hist[i,:] + p['fourier']['c'][m, i]*np.exp(1j*m*phi)
+        if self.boxfacecolor == 'inherit':
+            self.boxfacecolor = matplotlib.rcParams['axes.facecolor']
 
-    return phi, phi_hist
+        if self.boxedgecolor == 'inherit':
+            self.boxedgecolor = matplotlib.rcParams['axes.edgecolor']
+
+        self.fancybox = _val_or_rc(kwargs.pop("fancybox", None),
+                                              'legend.fancybox')
+
+        super().__init__(*args, **kwargs)
+
+
+    def draw(self, renderer):
+        super()._init()
+
+        # the following duplication of bits of super(renderer).draw is necessary to get the
+        # text bbox in the right place. Alternative is to actually call super(renderer).draw,
+        # but then we end up having to draw twice so that the contents is above the background.
+        pos = self.get_transform().transform((self.X, self.Y))
+        self.text.set_position(pos + self._text_shift())
+
+        if self.boxfacecolor is not None:
+            figure_inverse_trans = self.figure.transFigure.inverted()
+
+            # first find the bbox of the text, in figure coords
+            bbox = self.text.get_window_extent(renderer)
+            bbox = bbox.transformed(figure_inverse_trans)
+
+
+            # now find a bbox for the arrow, which is a subtle/annoying thing because the offsets and vertices
+            # are transformed differently
+            arrow_offsets = self.get_transform().transform(self.vector.get_offsets())
+            arrow_vertices =  self.Q.get_transform().transform(self.verts[0])
+
+            arrow_vertices_offset = arrow_offsets + arrow_vertices
+
+            x0y0 = arrow_vertices_offset.min(axis=0)
+            x1y1 = arrow_vertices_offset.max(axis=0)
+
+            x0y0 = figure_inverse_trans.transform(x0y0)
+            x1y1 = figure_inverse_trans.transform(x1y1)
+
+            # expand bbox to include all arrow_vertices:
+            bbox = matplotlib.transforms.Bbox.union([bbox, matplotlib.transforms.Bbox([x0y0, x1y1])])
+
+            # and, at last, we know the coordinates of the background that we need!
+
+            boxstyle = ("round,pad=0.02,rounding_size=0.02" if self.fancybox
+                        else "square,pad=0.02")
+
+            background = matplotlib.patches.FancyBboxPatch(
+                bbox.min, bbox.width, bbox.height, boxstyle=boxstyle,
+                fc=self.boxfacecolor, ec=self.boxedgecolor,
+                transform=self.figure.transFigure)
+
+            background.draw(renderer)
+        super().draw(renderer)
+
+
+def _test_quiverkey(scale=10.0, labelpos='E'):
+    """A simple test for the PynbodyQuiverKey class."""
+    import matplotlib.pyplot as p
+    import numpy as np
+    X, Y = np.meshgrid(np.arange(0, 2 * np.pi, .2), np.arange(0, 2 * np.pi, .2))
+    U = np.cos(X)
+    V = np.sin(Y)
+    fig, ax = p.subplots()
+    q = ax.quiver(X, Y, U, V)
+
+    qk = PynbodyQuiverKey(q, .5, .95, scale, "Quiver key", labelpos=labelpos,
+                          boxfacecolor='w', boxedgecolor='k', fancybox=True)
+    ax.add_artist(qk)
+    qk.set_zorder(5)
